Move CTR_DRBG emulator impl into separate module file
